Remove dt_3byte from DTYP_TO_SIZE
In IDA 7, dt_3byte and dt_byte both equal 0. This causes all byte
operands to be listed as size 3.

There is no `dt_3byte` in the ida_ua module, which is the module
containing all the `dt_*` fields in the idaapi reorganization HexRays
did.

When asked about dt_3byte not being in the ida_ua module, HexRays
responded with:

> Yes, we removed the 3byte operand type from IDA. It is not really used
> in any processor modules except one or two, where we reimplemented it as
> a custom data type. Feel free to remove all references to it from your
> source code.

Thus, this commit removes the usage of dt_3byte as per HexRay's
response.

temp